# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.ProductResponse:
    properties:
      category:
        type: string
      created_at:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      sku:
        type: string
      stock:
        type: integer
      updated_at:
        type: string
    type: object
...
  response.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Point of Sale (POS) application with multi-tenant support.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: POS Backend API
  version: "1.0"
paths:
...
  /products:
    get:
      description: Retrieve list of products (optionally filtered by tenant_id)
      parameters:
      - description: Tenant ID
        in: query
        name: tenant_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all products
      tags:
      - Product
    post:
      consumes:
      - multipart/form-data
      description: Create a new product (multipart/form-data). Include `image` file
        optionally.
      parameters:
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product SKU
        in: formData
        name: sku
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: string
      - description: Tenant ID
        in: formData
        name: tenant_id
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: Stock
        in: formData
        name: stock
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Product
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Product
    get:
      description: Retrieve a single product by its ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get product by ID
      tags:
      - Product
    put:
      consumes:
      - multipart/form-data
      description: Update an existing product (multipart/form-data). `image` optional.
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        type: string
      - description: Product name
        in: formData
        name: name
        required: true
        type: string
      - description: Product SKU
        in: formData
        name: sku
        required: true
        type: string
      - description: Category ID
        in: formData
        name: category_id
        required: true
        type: string
      - description: Tenant ID
        in: formData
        name: tenant_id
        required: true
        type: string
      - description: Price
        in: formData
        name: price
        required: true
        type: number
      - description: Stock
        in: formData
        name: stock
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Product
  /products/filter:
    get:
      description: Retrieve products using filters (name, sku, category_id) and pagination
      parameters:
      - description: Name
        in: query
        name: name
        type: string
      - description: SKU
        in: query
        name: sku
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: Tenant ID
        in: query
        name: tenant_id
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ProductResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get products with filter and pagination
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"